+++ src/main/java/com/robertx22/age_of_exile/database/registry/SlashRegistryType.java
@@ -13,6 +13,7 @@ import com.robertx22.age_of_exile.database.data.exile_effects.ExileEffect;
 import com.robertx22.age_of_exile.database.data.gear_slots.GearSlot;
 import com.robertx22.age_of_exile.database.data.gear_types.bases.BaseGearType;
 import com.robertx22.age_of_exile.database.data.gems.Gem;
+import com.robertx22.age_of_exile.database.data.groups.GearRarityGroup;
 import com.robertx22.age_of_exile.database.data.perks.Perk;
 import com.robertx22.age_of_exile.database.data.rarities.GearRarity;
 import com.robertx22.age_of_exile.database.data.rarities.MobRarity;
@@ -66,7 +67,8 @@ public enum SlashRegistryType {
     },
     SPELL("spells", 17, Spell.SERIALIZER, SyncTime.ON_LOGIN),
     PERK("perk", 18, Perk.SERIALIZER, SyncTime.ON_LOGIN),
-    SPELL_SCHOOL("spell_school", 19, SpellSchool.SERIALIZER, SyncTime.ON_LOGIN);
+    SPELL_SCHOOL("spell_school", 19, SpellSchool.SERIALIZER, SyncTime.ON_LOGIN),
+    GEAR_RARITY_GROUP("gear_rarity_group", 20, GearRarityGroup.SERIALIZER, SyncTime.ON_LOGIN);
 
     public String id;
     ISerializable ser;