+++ src/main/java/com/robertx22/age_of_exile/loot/blueprints/GearBlueprint.java
@@ -1,15 +1,12 @@
 package com.robertx22.age_of_exile.loot.blueprints;
 
+import com.robertx22.age_of_exile.database.data.rarities.GearRarity;
 import com.robertx22.age_of_exile.database.registry.RarityRegistryContainer;
 import com.robertx22.age_of_exile.database.registry.SlashRegistry;
 import com.robertx22.age_of_exile.loot.LootInfo;
-import com.robertx22.age_of_exile.loot.blueprints.bases.GearItemSlotPart;
-import com.robertx22.age_of_exile.loot.blueprints.bases.IsUniquePart;
-import com.robertx22.age_of_exile.loot.blueprints.bases.UnidentifiedPart;
-import com.robertx22.age_of_exile.loot.blueprints.bases.UniqueGearPart;
+import com.robertx22.age_of_exile.loot.blueprints.bases.*;
 import com.robertx22.age_of_exile.loot.generators.stack_changers.DamagedGear;
 import com.robertx22.age_of_exile.loot.generators.util.GearCreationUtils;
-import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.Rarity;
 import com.robertx22.age_of_exile.saveclasses.item_classes.GearItemData;
 import net.minecraft.item.ItemStack;
 
@@ -33,13 +30,15 @@ public class GearBlueprint extends ItemBlueprint {
         actionsAfterGeneration.add(DamagedGear.INSTANCE);
     }
 
+    public GearRarityPart rarity = new GearRarityPart(this);
+
     public GearItemSlotPart gearItemSlot = new GearItemSlotPart(this);
     public UnidentifiedPart unidentifiedPart = new UnidentifiedPart(this);
     public UniqueGearPart uniquePart = new UniqueGearPart(this);
     public IsUniquePart isUniquePart = new IsUniquePart(this);
 
     @Override
-    public RarityRegistryContainer<? extends Rarity> getRarityContainer() {
+    public RarityRegistryContainer<GearRarity> getRarityContainer() {
         return SlashRegistry.GearRarities();
     }