+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 + Reverted rarities to mine and slash ones. That means epic, legendary and mythic are back! I also added antique rarity.
 Antiques are basically easier to get but worse relics.
 + compatible item datapack customization fields are now optional and have defaults
++ compatible item now uses gear rarity group datapacks instead of min and max rarity rank fields. 
+The ranks were confusing so they're gone. Now each rarity can tell if some other rarity is a higher version of itself or not.
+The default if it's not inputted is all rarities besides unique.
 + durability tooltip now says unbrekable if item is unbreakable.
 + added more hp to npcs and less to animals
 + gear name tooltips are now on 2 lines sometimes to reduce tooltip length